+++ src/pages/tireDetailPage/tireFeedback.css
@@ -1,0 +1,104 @@
+.allRatingBlock{
+  display: flex;
+  gap: 20px;
+  margin-top: 20px;
+}
+
+.feedback {
+    border-radius: 10px;
+    width: 985px;
+    height: 232px;
+    display: flex;
+    justify-content: space-between;
+    background-color: #F79C1321;
+    padding: 10px;
+    flex-direction: column;
+  }
+  
+  .feedback__ratings {
+    display: flex;
+    justify-content: space-between;
+    width: 100%;
+  }
+  
+  .feedback__user {
+    font-family: Inter;
+    font-size: 18px;
+    font-weight: 600;
+    line-height: 21.78px;
+    text-align: left;
+  }
+  
+  .feedback__date {
+    font-family: Inter;
+    font-size: 15px;
+    font-weight: 400;
+    line-height: 18.15px;
+    text-align: left;
+  }
+  
+  .feedback__body {
+    flex-grow: 1;
+    margin-top: 10px;
+  }
+  
+  .feedback__text {
+    font-family: Inter;
+    font-size: 16px;
+    font-weight: 400;
+    line-height: 19.36px;
+    text-align: left;
+  }
+  
+  .feedback__marks {
+    display: flex;
+    justify-content: space-between;
+    width: 10%;
+  }
+  
+  .feedback__likes,
+  .feedback__dislikes {
+    margin-bottom: 10px;
+    text-align: center;
+    color: #8F8F8F;
+    font-family: Inter;
+    font-size: 16px;
+    font-weight: 400;
+    line-height: 19.36px;
+  }
+  
+  .feedback__like-img,
+  .feedback__dislike-img {
+    margin-right: 5px;
+  }
+  
+  .rating-container {
+    width: 315px;
+    height: 232px;
+    border: 1px solid #dbf800;
+    border-radius: 8px;
+      display: flex;
+  flex-direction: column;
+padding: 20px;
+  height: 100%;
+  justify-content: space-around;
+  }
+  
+  .feedback__extra-star {
+    margin-top: 10px;
+  }
+  
+  .feedback__extra-star-value {
+    display: flex;
+    align-items: center;
+    font-family: Inter;
+    font-size: 16px;
+    font-weight: 400;
+    line-height: 19.36px;
+    text-align: left;
+  }
+  .like_img,.dislike_img{
+    width: 16px;
+    height: 16px;
+    margin-right: 5px;
+  }
